apps/python_backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "denga-python-backend"
version = "0.1.0"
description = "FastAPI + worker backend for Denga"
requires-python = ">=3.12"
dependencies = [
  "bcrypt==4.2.1",
  "email-validator==2.2.0",
  "fastapi==0.115.12",
  "httpx==0.28.1",
  "passlib[bcrypt]==1.7.4",
  "psycopg[binary]==3.2.6",
  "pydantic-settings==2.8.1",
  "PyJWT==2.10.1",
  "python-dotenv==1.1.0",
  "sqlalchemy==2.0.40",
  "uvicorn[standard]==0.34.1",
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.5",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
